refactor: reorganize code to move out of root room folder
